requirements.txt
pygame-ce==2.5.6
numpy==2.3.4
requests==2.32.5
gif_pygame==1.2.0
packaging>=23.2
